Use SmartWeave::caller for reading caller identity
Instead of using Transaction::owner(), use SmartWeave::caller() when
reading smart contract caller identity. The difference here is that
the SmartWeave::caller will use smart contract's identity as the caller
when doing foreign contract calls. This is most of the time what is
desired so use it instead of the Transaction::owner().