Make wasm-bindgen optional
Validator app depends on validators contracts, but only needs the
access to type definitions. Including wasm-bindgen creates dependency
conflict between validator app and validators contract when the version
of the wasm-bindgen requires too old version of chrono crate.

This change is a quick hack to allow validator app to depend on
validators contract crate without pulling wasm-bindgen as a dependency.

While we need better split of the code, this allows us to move forward
for now without more extensive restructure of the code.